+++ frontend/src/lib/utils/eventUsageLogic.ts
@@ -41,7 +41,7 @@ import {
     isRetentionFilter,
     isStickinessFilter,
     isTrendsFilter,
-} from 'scenes/insights/utils/cleanFilters'
+} from 'scenes/insights/sharedUtils.ts'
 export enum DashboardEventSource {
     LongPress = 'long_press',
     MoreDropdown = 'more_dropdown',
