+++ frontend/src/scenes/funnels/funnelLogic.ts
@@ -53,7 +53,8 @@ import {
     generateBaselineConversionUrl,
 } from './funnelUtils'
 import { dashboardsModel } from '~/models/dashboardsModel'
-import { cleanFilters, isFunnelsFilter } from 'scenes/insights/utils/cleanFilters'
+import { cleanFilters } from 'scenes/insights/utils/cleanFilters'
+import { isFunnelsFilter } from 'scenes/insights/sharedUtils.ts'
 import { keyForInsightLogicProps } from 'scenes/insights/sharedUtils'
 import { teamLogic } from '../teamLogic'
 import { personPropertiesModel } from '~/models/personPropertiesModel'
