+++ frontend/src/scenes/insights/InsightDisplayConfig.tsx
@@ -23,7 +23,7 @@ import {
     isRetentionFilter,
     isStickinessFilter,
     isTrendsFilter,
-} from 'scenes/insights/utils/cleanFilters'
+} from 'scenes/insights/sharedUtils.ts'
 
 interface InsightDisplayConfigProps {
     filters: FilterType
