+++ frontend/src/scenes/paths/pathsLogic.ts
@@ -4,7 +4,8 @@ import { insightLogic } from 'scenes/insights/insightLogic'
 import type { pathsLogicType } from './pathsLogicType'
 import { InsightLogicProps, FilterType, PathType, PropertyFilter, InsightType, PathsFilterType } from '~/types'
 import { keyForInsightLogicProps } from 'scenes/insights/sharedUtils'
-import { cleanFilters, isPathsFilter } from 'scenes/insights/utils/cleanFilters'
+import { cleanFilters } from 'scenes/insights/utils/cleanFilters'
+import { isPathsFilter } from 'scenes/insights/sharedUtils.ts'
 import { urls } from 'scenes/urls'
 import { TaxonomicFilterGroupType } from 'lib/components/TaxonomicFilter/types'
 import { openPersonsModal } from 'scenes/trends/persons-modal/PersonsModal'
