+++ frontend/src/scenes/retention/retentionTableLogic.ts
@@ -5,7 +5,8 @@ import { RETENTION_FIRST_TIME, RETENTION_RECURRING } from 'lib/constants'
 import { range, toParams } from 'lib/utils'
 import { insightLogic } from 'scenes/insights/insightLogic'
 import { keyForInsightLogicProps } from 'scenes/insights/sharedUtils'
-import { cleanFilters, isRetentionFilter } from 'scenes/insights/utils/cleanFilters'
+import { cleanFilters } from 'scenes/insights/utils/cleanFilters'
+import { isRetentionFilter } from 'scenes/insights/sharedUtils.ts'
 import { RetentionTablePayload, RetentionTablePeoplePayload, RetentionTrendPayload } from 'scenes/retention/types'
 import { actionsModel } from '~/models/actionsModel'
 import { Noun, groupsModel } from '~/models/groupsModel'
